Fix withdrawal happy part test
